Replaces main app logo
HACK: The main app logo is currently composed of the graphic component returned by ReplayWeb.page which is then assigned a different variable in ArchiveWeb.page with `get mainLogo`.  The text is then appended by ArchiveWeb.page.

Because the text and icon are now located in the same file for both apps, this behavior is worth cleaning up once both are up to date and using TypeScript.  For now, this returns nothing as the `mainLogo` value and appends the whole new logo lockup instead of just the text.